day9 further cleanup
Using frozenset since that like tuple is immutable and can be used as a
dictionary key (since it can be hashed). The upside is that the hash
does not depend on order which is what we want.

permutations by default gives all permutations of the full length so no
need for the second parameter.