day9 added path namespace
The __new__ function is not really needed, but it makes the
"initialization" nicer.

This was a bit tricky to get right since overriding __init__ in
immutable classes is not what you expect. __new__ is a static method
that is called before __init__ and is the right place to do stuff like
this.