2019 Added a cli
Will continue to copy-pase the current code to the next days folder for
now. This does not related to the day4 task (which have not been
released yet).

The cli might be a misguided effort but it is my try at getting a nice
step through debugger which prints info at each step.