2019 day5 computer arg cleanup
Most arguments should follow the param_mode, and those you get for free
correctly as parameters to functions using the instruction decorator.

However some instructions write to memory. The day5 instructions state:

    "Parameters that an instruction writes to will never be in immediate
     mode."

This means special care needs to be taken for these. Hence the ability
to fetch the raw parameters so that instructions that need those can do
so.