2020 day18 part2 fixed
Brackets handling is still ugly, in fact a lot of this is ugly.
But now the last bug was fixed.

    (9 * 1) + 1

This brackets were previously overriden by + so the evaluation order
became:

    9 * (1 + 1)

I added a brackets property and made sure the rearanging to make + have
does not split up brackets on the left.